Fixes the new quest system page
After the super smooth switchover to a new server, to fix performance the quest
overview was moved onto it's own page, and as a result it was having table rows
set to `display: flex;`. The overly large selector is needed (there may well be
a better way to resolve this, SCSS is not my forte) to override the `display: flex`.

Trying to use !important broke the hide/show functionality since it just sets `display: none`
on the tr element.